clients/requirements.txt
requests
click
jsonschema
pyjwt
faker
aiohttp